DEV-3317 remove PurpleSomaticVariant and use PurpleVariant instead
