+++ orange-datamodel/pom.xml
@@ -42,5 +42,9 @@
             <artifactId>junit</artifactId>
             <scope>test</scope>
         </dependency>
+        <dependency>
+            <groupId>com.hartwig</groupId>
+            <artifactId>hmf-common</artifactId>
+        </dependency>
     </dependencies>
 </project>
