--- orange/src/main/java/com/hartwig/hmftools/orange/algo/purple/PurpleSomaticVariant.java
@@ -1,126 +1,0 @@
-package com.hartwig.hmftools.orange.algo.purple;
-
-import java.util.List;
-
-import com.hartwig.hmftools.common.genotype.GenotypeStatus;
-import com.hartwig.hmftools.common.purple.GermlineStatus;
-import com.hartwig.hmftools.common.variant.AllelicDepth;
-import com.hartwig.hmftools.common.variant.CodingEffect;
-import com.hartwig.hmftools.common.variant.Hotspot;
-import com.hartwig.hmftools.common.variant.SomaticLikelihood;
-import com.hartwig.hmftools.common.variant.SomaticVariantFactory;
-import com.hartwig.hmftools.common.variant.Variant;
-import com.hartwig.hmftools.common.variant.VariantTier;
-import com.hartwig.hmftools.common.variant.VariantType;
-import com.hartwig.hmftools.common.variant.impact.VariantTranscriptImpact;
-
-import org.jetbrains.annotations.NotNull;
-import org.jetbrains.annotations.Nullable;
-
-public interface PurpleSomaticVariant extends Variant {
-
-    double qual();
-
-    @NotNull
-    String filter();
-
-    int genesAffected();
-
-    boolean spliceRegion();
-
-    @NotNull
-    String otherReportedEffects();
-
-    @NotNull
-    CodingEffect worstCodingEffect();
-
-    @NotNull
-    Hotspot hotspot();
-
-    boolean recovered();
-
-    default boolean isHotspot() {
-        return hotspot() == Hotspot.HOTSPOT;
-    }
-
-    double mappability();
-
-    default boolean isFiltered() {
-        return !filter().equals(SomaticVariantFactory.PASS_FILTER);
-    }
-
-    default boolean isSnp() {
-        return type() == VariantType.SNP;
-    }
-
-    double adjustedCopyNumber();
-
-    double adjustedVAF();
-
-    double minorAlleleCopyNumber();
-
-    double variantCopyNumber();
-
-    boolean biallelic();
-
-    boolean reported();
-
-    @NotNull
-    GenotypeStatus genotypeStatus();
-
-    @NotNull
-    GermlineStatus germlineStatus();
-
-    @NotNull
-    String trinucleotideContext();
-
-    @NotNull
-    String microhomology();
-
-    @NotNull
-    String repeatSequence();
-
-    int repeatCount();
-
-    @NotNull
-    String kataegis();
-
-    @NotNull
-    VariantTier tier();
-
-    double subclonalLikelihood();
-
-    default double clonalLikelihood() {
-        return 1 - subclonalLikelihood();
-    }
-
-    @Nullable
-    AllelicDepth rnaDepth();
-
-    @Nullable
-    AllelicDepth referenceDepth();
-
-    @Nullable
-    List<Integer> localPhaseSets();
-
-    default Integer topLocalPhaseSet() {
-        return localPhaseSets() != null && !localPhaseSets().isEmpty() ? localPhaseSets().get(0) : null;
-    }
-
-    default String localPhaseSetsStr() {
-        return SomaticVariantFactory.localPhaseSetsStr(localPhaseSets());
-    }
-
-    default boolean hasLocalPhaseSets() {
-        return localPhaseSets() != null && !localPhaseSets().isEmpty();
-    }
-
-    String clinvarInfo();
-
-    double gnomadFrequency();
-
-    SomaticLikelihood somaticLikelihood();
-
-    List<VariantTranscriptImpact> variantTranscriptImpacts();
-
-}
